+++ lib/models/store/app_state.dart
@@ -33,7 +33,7 @@ class AppState with _$AppState {
     @Default(false) bool shuffle,
     @Default(Repeat.none) Repeat repeat,
     @Default(BoxFit.contain) BoxFit fit,
-    @Default(100) int volume,
+    @Default(80) int volume,
     @Default(false) bool isMuted,
     @Default(ThemeMode.system) ThemeMode themeMode,
     @Default('none') String preferedSubtitleLanguage,