+++ lib/store/use_app_store.dart
@@ -62,6 +62,26 @@ class AppStore extends PersistentStore<AppState> {
     await save(state);
   }
 
+  Future<void> updateVolume(int volume) async {
+    set(state.copyWith(
+        volume: volume < 0
+            ? 0
+            : volume > 100
+                ? 100
+                : volume));
+    await save(state);
+  }
+
+  Future<void> updateMute(bool isMuted) async {
+    set(state.copyWith(isMuted: isMuted));
+    await save(state);
+  }
+
+  Future<void> toggleMute() async {
+    set(state.copyWith(isMuted: !state.isMuted));
+    save(state);
+  }
+
   Future<void> updateThemeMode(ThemeMode themeMode) async {
     set(state.copyWith(themeMode: themeMode));
     await save(state);